refactor: unit management around squads encapsulated in SquadManager fix: consider hostile static defense in fight simulation
